+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.12] 2018-07-23
+### Changed
+- Added a parent frame to the party frames handling hiding in raids, leaving the party frames own visibility drivers untouched.
+- Made a larger part of the player energy crystal / mana orb area right-clickable in addition to the player health bar.
+- Made the actionbutton mask texture far more perfect in shape to make it work better with actionbutton fade-outs.  
+
+### Fixed 
+- Fixed an issue that caused the Artifact Power bar to not properly be displayed on the first login. 
+
 ## [1.0.11] 2018-07-23
 ### Fixed
 - Fixed an issue that caused action buttons containing spell flyouts to become tainted and prevent casting.